+++ src/cli.spec.ts
@@ -14,13 +14,13 @@ describe('Test generateRandomString CLI', () => {
         process.argv = ['', ''];
 
         // Run command.
-        expect(() => require('./cli')).toThrow('A number must be given as command parameter.');
+        expect(() => require('./cli')).toThrow('A number must be given as the first command parameter.');
     });
 
     test.each([
 
         ['-5', 'Length must be a positive number.'],
-        ['Not a Number', 'The given command parameter is not a number.'],
+        ['Not a Number', 'The first command parameter is not a number.'],
 
     ])('Invalid command parameters', (input, error) => {
 
@@ -31,21 +31,16 @@ describe('Test generateRandomString CLI', () => {
         expect(() => require('./cli')).toThrow(error);
     });
 
-    test('Calling the command correctly', () => {
+    test.each([
+        [['', '', '5']],
+        [['', '', '5', undefined]],
+        [['', '', '5', '']],
+        [['', '', '5', 'Something else']],
+        [['', '', '5', '123']],
+    ])('Correct calls', (input) => {
 
         // Set command argument.
-        process.argv = ['', '', '5'];
-
-        // Run command.
-        require('./cli');
-        // tslint:disable-next-line:no-console
-        expect(console.log).toHaveBeenCalledTimes(1);
-    });
-
-    test('Giving extra arguments', () => {
-
-        // Set command argument.
-        process.argv = ['', '', '5', 'Something else'];
+        process.argv = input as any;
 
         // Run command.
         require('./cli');